+++ Sources/Epic/Protocols/ThrottleProtocol.swift
@@ -1,0 +1,11 @@
+/**
+ Protocols provide an easy interface to swap Epic objects with your own.
+ */
+
+import Foundation
+
+protocol ThrottleProtocol {
+    init(interval: TimeInterval)
+
+    func throttle(block: (() -> Void)?)
+}