+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## 0.1.x Releases
 - 0.1.0 ~ Initial draft
-- 0.1.1 ~ Documentation and internal renamings
+- 0.1.1_Swift5 ~ Additional support for Swift 5
+- 0.2.0 ~ Support for Swift Package Manager and ceasing support for Carthage
+
 ### Added
 - Throttle
 - Generator
-- Message Bus + Topic + Broker
+- Event Bus: Message Bus, Topic, Broker